Refactored handleViewContent so it can be reused
ref https://linear.app/ghost/issue/AP-540/clicking-comment-icon-on-posts-and-likes-tabs-of-your-profile-doesnt

- We want to open posts in the drawer from multiple views (Inbox, Profile etc.) and this change allows us to do so, while simplifying the code
